Moved inspector to a new file.
Now it seems to properly find all the files, either Go and non-go.

Logs show that non-go files are also instrumented, but by some reason
they are not properly forwarding eBPF data.
