+++ test/integration/docker-compose-multiexec.yml
@@ -18,9 +18,6 @@ services:
     image: hatest-testserver-python
     ports:
       - "8081:8080"
-    depends_on:
-      otelcol:
-        condition: service_started
 
   autoinstrumenter:
     build:
@@ -51,6 +48,9 @@ services:
       INTERNAL_METRICS_PROMETHEUS_PORT: 8999
     ports:
       - "8999:8999" # Prometheus scrape port, if enabled via config
+    depends_on:
+      testserver-python:
+        condition: service_started
 
   # OpenTelemetry Collector for Metrics. For Traces, we use directly Jaeger
   otelcol:
